# api-contract.yaml
openapi: 3.0.3
info:
  title: Debt Management System API
  description: >
    Автоматизированная система управления дебиторской задолженностью
    с модулем омниканальных уведомлений (SMS / WhatsApp / Telegram / Email).
  version: 1.0.0
  contact:
    name: Dev Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.debtmanager.example.com/v1
    description: Production
...
paths:
...
  /auth/login/:
    post:
      tags: [Auth]
      summary: Получить JWT-токены
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: agent_ivanov
                password:
                  type: string
                  format: password
                  example: "s3cr3t"
      responses:
        "200":
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh/:
    post:
      tags: [Auth]
      summary: Обновить access-токен
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh]
              properties:
                refresh:
                  type: string
      responses:
        "200":
          description: Новый access-токен
          content:
            application/json:
              schema:
                type: object
                properties:
                  access:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout/:
    post:
      tags: [Auth]
      summary: Инвалидировать refresh-токен
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh]
              properties:
                refresh:
                  type: string
      responses:
        "204":
          description: Выход выполнен

  # ── USERS ─────────────────────────────
...
  /debtors/:
    get:
      tags: [Debtors]
      summary: Список должников
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: search
          in: query
          schema:
            type: string
          description: Поиск по ФИО или телефону
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDebtors"
    post:
      tags: [Debtors]
      summary: Создать должника
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtorCreate"
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debtor"
  /debtors/import/:
    post:
      tags: [Debtors]
      summary: Массовый импорт из CSV/Excel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "202":
          description: Импорт принят в очередь
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  message:
                    type: string
                    example: "Import queued. 120 rows detected."
  /debtors/{id}/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    get:
      tags: [Debtors]
      summary: Получить должника
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debtor"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Debtors]
      summary: Обновить данные должника
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtorUpdate"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debtor"
    delete:
      tags: [Debtors]
      summary: Удалить должника
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Удалено

  # ── DEBT CASES ────────────────────────
  /debt-cases/:
    get:
      tags: [DebtCases]
      summary: Список дел о задолженности
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_progress, promised, closed, overdue]
        - name: dpd_min
          in: query
          schema:
            type: integer
          description: Фильтр по минимальному DPD
        - name: dpd_max
          in: query
          schema:
            type: integer
        - name: assigned_agent_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDebtCases"
    post:
      tags: [DebtCases]
      summary: Создать дело
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtCaseCreate"
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtCase"
  /debt-cases/{id}/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    get:
      tags: [DebtCases]
      summary: Получить дело
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtCase"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [DebtCases]
      summary: Обновить дело (статус, агент)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtCaseUpdate"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtCase"
  /debt-cases/{id}/dpd-history/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    get:
      tags: [DebtCases]
      summary: История DPD по делу
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DPDSnapshot"

  # ── NOTIFICATIONS ─────────────────────
...
  /integrations/:
    get:
      tags: [Integrations]
      summary: Список конфигураций каналов
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntegrationConfig"
    post:
      tags: [Integrations]
      summary: Добавить конфигурацию канала
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationConfigCreate"
      responses:
        "201":
          description: Создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationConfig"
  /integrations/{id}/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    patch:
      tags: [Integrations]
      summary: Обновить конфигурацию
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationConfigCreate"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationConfig"
    delete:
      tags: [Integrations]
      summary: Удалить конфигурацию
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Удалено
  /integrations/{id}/test/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    post:
      tags: [Integrations]
      summary: Тест соединения с каналом
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Результат теста
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "Connection to WhatsApp Ch2D OK"

# ─────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UUIDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
    Unauthorized:
      description: Не аутентифицирован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Not found."
    TokenPair:
      type: object
      properties:
        access:
          type: string
        refresh:
          type: string

    # ── USER ──
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        created_at:
          type: string
          format: date-time
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: agent
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
...
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codename:
          type: string
          example: can_send_notification
        description:
          type: string

    # ── DEBTOR ──
    Debtor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          example: Иванов Иван Иванович
        phone:
          type: string
          example: "+996700123456"
        email:
          type: string
          format: email
          nullable: true
        whatsapp_number:
          type: string
          nullable: true
        telegram_id:
          type: string
          nullable: true
    DebtorCreate:
      type: object
      required: [full_name, phone]
      properties:
        full_name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        whatsapp_number:
          type: string
        telegram_id:
          type: string
    DebtorUpdate:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        whatsapp_number:
          type: string
        telegram_id:
          type: string
    PaginatedDebtors:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/Debtor"

    # ── DEBT CASE ──
    DebtCase:
      type: object
      properties:
        id:
          type: string
          format: uuid
        debtor:
          $ref: "#/components/schemas/Debtor"
        assigned_agent:
          $ref: "#/components/schemas/User"
          nullable: true
        amount:
          type: number
          format: float
          example: 15000.00
        due_date:
          type: string
          format: date
        dpd:
          type: integer
          description: Days Past Due
          example: 7
        status:
          type: string
          enum: [new, in_progress, promised, closed, overdue]
        created_at:
          type: string
          format: date-time
    DebtCaseCreate:
      type: object
      required: [debtor_id, amount, due_date]
      properties:
        debtor_id:
          type: string
          format: uuid
        assigned_agent_id:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
          format: float
        due_date:
          type: string
          format: date
    DebtCaseUpdate:
      type: object
      properties:
        assigned_agent_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [new, in_progress, promised, closed, overdue]
    DPDSnapshot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        debt_case_id:
          type: string
          format: uuid
        dpd_value:
          type: integer
        snapshot_date:
          type: string
          format: date
    PaginatedDebtCases:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/DebtCase"

    # ── NOTIFICATION TEMPLATE ──
...
    IntegrationConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        channel:
          type: string
          enum: [whatsapp, sms, telegram, email]
        provider:
          type: string
          example: ch2d
        webhook_url:
          type: string
          format: uri
          nullable: true
        is_active:
          type: boolean
    IntegrationConfigCreate:
      type: object
      required: [channel, provider, api_key]
      properties:
        channel:
          type: string
          enum: [whatsapp, sms, telegram, email]
        provider:
          type: string
          example: ch2d
        api_key:
          type: string
          format: password
          writeOnly: true
        webhook_url:
          type: string
          format: uri
        is_active:
          type: boolean
          default: true

    # ── SHARED ──
    Pagination:
      type: object
      properties:
        count:
          type: integer
          description: Всего объектов
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true